\name{TickExec-package}
\alias{TickExec-package}
\alias{TickExec}
\docType{package}
\title{Execution Functions for Tick Data Back Test}
\description{
Functions to execute orders in backtesting using tick data. A testing platform was established by the four major execution functions, namely 'LimitBuy', 'LimitSell', 'MarketSell' and 'MarketSell', which enclosed all tedious aspects (such as queueing for order executions and calculate actual executed volumes) for order execution using tick data. Such that one can focus on the logic of strategies, rather than its execution.
}
\details{
  \tabular{ll}{
  Package: \tab TickExec\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2015-04-20\cr
  License: \tab PGL-3\cr
  }
}
\author{
  HKUST.
  Maintainer: SONG Yang <ysongad@connect.ust.hk>
}

\keyword{ Tick, Back test, Execution }